Fix anchors hidden behind fixed header
Jumping to a section via #anchor didn't work properly as the headline was hidden by the fixed header. Normally, this would be fixed by the :target CSS rule. But this doesn't work with the id on the headline.

Moving the id to the section itself fixes this. It also makes the sections of an article easier to browse in the source code view of a browser.